+++ src/app/commands/admin/Account.ts
@@ -44,12 +44,10 @@ export default class Account extends SlashCommand {
 		if (ctx.user.bot) return;
 		const { guildMember } = await ServiceUtils.getGuildAndMember(ctx);
 		try {
-			await VerifyTwitter(guildMember);
-			await ctx.send({ content: 'Please check your DMs.', ephemeral: true });
+			await VerifyTwitter(ctx, guildMember);
 		} catch (e) {
 			LogUtils.logError('failed to verify user', e, guildMember.guild.id);
-			await ctx.send({ content: 'Welp, something is definitely broken. I would blame you, but I know better. I\'ll let my devs ' +
-				'know something is wrong.', ephemeral: true });
+			await ServiceUtils.sendOutErrorMessage(ctx);
 		}
 	}
 }